Renamed algorithm utils. Refactored static imports. Made algorithm readable
