+++ src/main/java/pl/edu/agh/firecell/engine/algorithm/BasicAlgorithm.java
@@ -16,11 +16,6 @@ public class BasicAlgorithm implements Algorithm {
     private final FirePropagator firePropagator;
     private final SmokePropagator smokePropagator;
     private final DiffusionGenerator diffusionGenerator;
-    public static final double CONVECTION_COEFFICIENT = 1;
-    // should be dependent on the material in the future
-    public static final double CONDUCTIVITY_COEFFICIENT = 0.2;
-    public static final int MAX_BURNING_TIME = 50;
-
 
     public BasicAlgorithm(double deltaTime) throws ConductionCoefficientException {
         this.temperaturePropagator = new TemperaturePropagator(deltaTime);
@@ -33,26 +28,18 @@ public class BasicAlgorithm implements Algorithm {
     public Cell compute(State oldState, Vector3i cellIndex) {
 
         Cell oldCell = oldState.getCell(cellIndex);
-        double newTemperature = oldCell.temperature();
-        int newBurningTime;
-        int newRemainingFirePillar;
-        boolean newFlammable;
-        // NOTE: Following method calls are dependent on each other, the order matters.
-        // Conduction
-        newTemperature = temperaturePropagator.computeConduction(oldState, cellIndex, newTemperature);
-        // Convection
-        if(oldCell.material().getMatterState().equals(MatterState.FLUID))
-            newTemperature = temperaturePropagator.computeConvection(oldState, cellIndex, newTemperature);
-        // Smoke update
-        double newSmokeIndicator = smokePropagator.computeNewSmokeIndicator(oldState, cellIndex, oldCell);
-        // Fire status update
-        newRemainingFirePillar = firePropagator.computeFirePillar(oldState, oldCell, cellIndex, oldCell.remainingFirePillar());
-        newBurningTime = firePropagator.computeBurningTime(oldState, oldCell, cellIndex, newTemperature);
-        newFlammable = firePropagator.computeNewFlammable(oldCell, newBurningTime);
-        newTemperature = temperaturePropagator.updateTemperatureBasedOnFire(oldCell, newTemperature);
 
-        // Diffusion update
+        double newTemperature = temperaturePropagator.computeConduction(oldState, cellIndex, oldCell.temperature());
+        if (oldCell.material().getMatterState().equals(MatterState.FLUID))
+            newTemperature = temperaturePropagator.computeConvection(oldState, cellIndex, newTemperature);
+
+        double newSmokeIndicator = smokePropagator.computeNewSmokeIndicator(oldState, cellIndex, oldCell);
         diffusionGenerator.smokeUpdate(oldState, cellIndex);
+
+        int newRemainingFirePillar = firePropagator.computeFirePillar(oldState, oldCell, cellIndex, oldCell.remainingFirePillar());
+        int newBurningTime = firePropagator.computeBurningTime(oldState, oldCell, cellIndex, newTemperature);
+        boolean newFlammable = firePropagator.computeNewFlammable(oldCell, newBurningTime);
+        newTemperature = temperaturePropagator.updateTemperatureBasedOnFire(oldCell, newTemperature);
 
         return new Cell(
                 newTemperature,
