+++ src/main/java/pl/edu/agh/firecell/engine/algorithm/TemperaturePropagator.java
@@ -8,12 +8,10 @@ import pl.edu.agh.firecell.model.State;
 import pl.edu.agh.firecell.model.util.NeighbourUtils;
 
 
-import static pl.edu.agh.firecell.engine.algorithm.BasicAlgorithm.*;
-import static pl.edu.agh.firecell.model.util.HelpfulFunctions.*;
-
 public class TemperaturePropagator {
 
     private static final double BURNING_TEMPERATURE_COEFFICIENT = 0.3;
+    private static final double CONVECTION_COEFFICIENT = 1;
 
     private final double deltaTime;
     private final MaterialConductionMap materialConductionMap;
@@ -57,7 +55,8 @@ public class TemperaturePropagator {
     }
 
     private double computeConductivity(Cell former, Cell middle, Cell latter, double conductivityCoeFormer, double conductivityCoeFurther) {
-        return -(conductivityCoeFormer * tempDiff(middle, former) + conductivityCoeFurther * tempDiff(middle, latter));
+        return -(conductivityCoeFormer * AlgorithmUtils.tempDiff(middle, former)
+                        + conductivityCoeFurther * AlgorithmUtils.tempDiff(middle, latter));
     }
 
     public double computeConvection(State oldState, Vector3i cellIndex, double currentTemperature) {
@@ -67,21 +66,21 @@ public class TemperaturePropagator {
         try {
             Cell cellUnder = oldState.getCell(NeighbourUtils.down(cellIndex));
             if (cellUnder.isFluid() && cellUnder.temperature() > oldCell.temperature())
-                temperatureDifference += CONVECTION_COEFFICIENT * tempDiffAbs(oldCell, cellUnder);
+                temperatureDifference += CONVECTION_COEFFICIENT * AlgorithmUtils.tempDiffAbs(oldCell, cellUnder);
         } catch (IndexOutOfBoundsException ignored) {
         }
 
         try {
             Cell cellAbove = oldState.getCell(NeighbourUtils.up(cellIndex));
             if (cellAbove.isFluid() && cellAbove.temperature() < oldCell.temperature())
-                temperatureDifference -= CONVECTION_COEFFICIENT * tempDiffAbs(oldCell, cellAbove);
+                temperatureDifference -= CONVECTION_COEFFICIENT * AlgorithmUtils.tempDiffAbs(oldCell, cellAbove);
         } catch (IndexOutOfBoundsException ignored) {
         }
         return currentTemperature + deltaTime * temperatureDifference;
     }
 
     public double updateTemperatureBasedOnFire(Cell oldCell, double currentTemperature) {
-        if (!isCellBurning(oldCell))
+        if (!AlgorithmUtils.isCellBurning(oldCell))
             return currentTemperature;
 
         double diffToMaterialBurningTemperature = oldCell.material().getBurningTemperature() - currentTemperature;
